requirements.txt
FastAPI
uvicorn
python-dotenv
aioimaplib
beautifulsoup4
email-validator
pydantic
httpx
jinja2
firebase-admin